Test : ServiceLayer변경에 따른 테스트케이스 수정
